requirements.txt
fastapi==0.120.1
uvicorn==0.38.0
kerykeion==5.1.1
requests==2.32.5
python-dotenv==1.2.1
python-multipart==0.0.20

# Performance optimization dependencies
redis==7.0.1
aiohttp==3.13.2
slowapi==0.1.9
httpx==0.28.1
aioredis==2.0.1
